fix: fixing pages paths and name pages corrections
