+++ frontend/package.json
@@ -16,6 +16,7 @@
     "axios": "^1.7.7",
     "leaflet": "^1.9.4",
     "lucide-react": "^0.460.0",
+    "next": "^15.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
