+++ frontend/src/AppRouter.tsx
@@ -1,23 +1,18 @@
 // src/AppRouter.tsx
-import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import NavBar from './components/navbar';
-import Home from './pages/Home';
-import Obras from './pages/Obras';
-import Mapas from './pages/Mapas';
+// import NavBar from './components/navbar';
 import Noticias from './pages/Noticias';
-import Login from './pages/Auth/Login';
-import Register from './pages/Auth/Register';
+import Menu from './pages/Menu';
+import Login from './pages/Login';
+import Register from './pages/Register';
 import Footer from './components/footer';
 
 const AppRouter = () => {
   return (
     <>
-      <NavBar />
+      {/* <NavBar /> */}
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/obras" element={<Obras />} />
-        <Route path="/mapa" element={<Mapas />} />
+        <Route path="/" element={<Menu />} />
         <Route path="/noticias" element={<Noticias />} />
         <Route path="/login" element={<Login />} />
         <Route path="/register" element={<Register />} />
