--- frontend/src/pages/Mapas.css
@@ -1,10 +1,0 @@
-.custom-marker {
-    filter: drop-shadow(2px 4px 6px black) brightness(1.2);
-  }
-  
-  .leaflet-container {
-    width: 100%;
-    height: 100%;
-  }
-  
-
