--- frontend/src/pages/Obras.tsx
@@ -1,8 +1,0 @@
-// src/pages/Obras.tsx
-import React from 'react';
-
-const Obras = () => {
-  return <div>Obras Page</div>;
-};
-
-export default Obras;
