+++ frontend/src/pages/Register.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useAuth } from '../../Context/AuthContext'
+import { useAuth } from '../Context/AuthContext'
 import { useNavigate } from 'react-router-dom';
 
 export default function Register() {
